fix: merge login, logout, register into one file
